requirements.txt
fastapi
sqlalchemy
psycopg2
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
uvicorn